pyproject.toml
[project]
name = "repofellow"
version = "0.1.0"
description = "AI-powered code analysis and modification tool with CLI interface"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
    "graphviz>=0.20.3",
    "jira>=3.8.0",
    "langchain-community>=0.3.14",
    "langchain-core>=0.3.29",
    "langchain-openai>=0.2.14",
    "langgraph>=0.2.61",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "langgraph-cli>=0.1.65",
    "langgraph-sdk>=0.1.48",
    "langsmith>=0.2.10",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "notebook>=7.3.2",
    "pathlib>=1.0.1",
    "pygithub>=2.5.0",
    "pylint>=3.3.3",
    "requests>=2.32.3",
    "rich>=13.7.0",
    "setuptools>=75.8.0",
    "tavily-python>=0.5.0",
    "tomli>=2.2.1",
    "trustcall>=0.0.26",
    "typer>=0.15.1",
    "wikipedia>=1.4.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/repofellow"
Documentation = "https://github.com/yourusername/repofellow#readme"
Repository = "https://github.com/yourusername/repofellow.git"
Issues = "https://github.com/yourusername/repofellow/issues"

[project.scripts]
repofellow = "repofellow.cli.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
